Final Commit BOYS
The final commit because i think everything is in place now.  Death Screen and Win Screens are in place.  Game restarts when the player dies and presses 'R'.  Theres an actual win condition which is find the raptor egg so that you can access the portal and go back to your time period.